requirements/dev.txt
# --- Development Dependencies ---
-r base.txt

# --- Dev & Quality Tools ---
black==24.4.2
isort==5.12.0
flake8==7.0.0
mypy==1.11.0
django-stubs==4.2.6
pre-commit==3.6.0
pytest==8.3.2
pytest-django==4.9.0
coverage==7.6.0
